Extract shared CapturingLogger test double from RedisCommandBudgetTests/RedisMultiplexerFactoryTests
Both test classes had a byte-for-byte identical ILogger test double. Extracted to a single
internal CapturingLogger in the test project so the two suites share one implementation
instead of maintaining duplicates.

Addresses review feedback on PR #1984.
